Fix all errors. Blink tested and working.

+++ RPI2_GPIO_C/include/gpio.h
@@ -4,8 +4,8 @@
 int pinMode(int pin, int mode);
 int digitalWrite(int pin, int value);
 int digitalRead(int pin);
-int blink(int pin, int freq, int duration);
-void delay(int dly);
+int blink(int pin, double freq, double duration);
+void delay(double dly);
 int unexportPin(int pin);
 
 #endif
